+++ ext/Server/__init__.lua
@@ -1,60 +1,4 @@
-local playerDamage
-local currentHiddenPlayer
-local playerDamageTable = {}
-
--- Save Damage done to Hidden in Table
-Hooks:Install('Soldier:Damage', 1, function(hook, soldier, info, giverInfo)
-	giverInfo = DamageGiverInfo(giverInfo)
-	--exclude hidden attacks, suicide ...
-	if giverInfo.giver ~= currentHiddenPlayer and giverInfo.damageType ~= DamageType.Suicide then
-		--print(DamageInfo(info).damage .. "dmg done to hidden from " .. giverInfo.giver.name)
-		for p,_ in pairs(playerDamageTable) do
-			if p == giverInfo.giver then
-				playerDamageTable[p] = playerDamageTable[p] + DamageInfo(info).damage
-				return
-			end
-		end
-		playerDamageTable[giverInfo.giver] = DamageInfo(info).damage
-	end
-end)
-
-function startRound()
-	if next(playerDamageTable) == nil then
-		--empty dmgTable so make Hidden random
-		playersOnServer = PlayerManager:GetPlayers()
-		currentHiddenPlayer = playersOnServer[MathUtils:GetRandomInt(1, PlayerManager:GetPlayerCount())]
-
-	else
-		local probabilityTable = {}
-		local totalDmg
-		for p, dmg in pairs(playerDamageTable) do
-			totalDmg = totalDmg + dmg
-		end
-		for p, dmg in pairs(playerDamageTable) do
-			probabilityTable[p] = dmg/totalDmg
-		end
-		currentHiddenPlayer = GetWeightedRandomKey(playerDamageTable)
-	end
-end
-
-function GetWeightedRandomKey( tab )
-	local sum = 0
-
-	for _, chance in pairs( tab ) do
-		sum = sum + chance
-	end
-
-	local select = MathUtils:GetRandom(0,1) * sum
-
-	for key, chance in pairs( tab ) do
-		select = select - chance
-		if select < 0 then return key end
-	end
-end
-
-function selectHidden()
-		local players = PlayerManager:GetPlayers()
-end
+require('selecthidden')
 
 function SpawnHidden(player)
 	if player == nil then
@@ -141,7 +85,11 @@ Events:Subscribe('Player:Chat', function(player, recipientMask, message)
 		ChatManager:Yell(testmsg, 4)
 	elseif message == "start" then
 		startRound()
-		ChatManager:Yell("Current hidden: " .. currentHiddenPlayer.name, 4)
+		ChatManager:Yell("Current hidden: " .. getHiddenPlayer().name, 4)
+	elseif message == "test1" then
+		player.teamId = TeamId.Team1
+	elseif message == "test2" then
+		player.teamId = TeamId.Team2
 	else
 		ChatManager:Yell(message, 3)
 	end
